* cleaup lastNotifiedLogger

+++ lastNotifiedLogger.py
@@ -1,19 +1,7 @@
 __author__ = 'fox'
-import get_update as gu
-import sys
 import os
 import datetime
 
-try:
-    from PyQt5 import QtGui, QtCore, QObject
-    from PyQt5.QtWidgets import QWidget, QApplication, QCursor, QFileDialog, QCursor, QIcon, QMessageBox
-    from PyQt5.QtCore import Qt, QThread, QObject
-    from PyQt4.QtCore import pyqtSignal
-except ImportError:
-    from PyQt4 import QtGui, QtCore
-    from PyQt4.QtGui import QWidget, QApplication, QCursor, QFileDialog, QCursor, QIcon, QMessageBox
-    from PyQt4.QtCore import Qt, QThread, QObject
-    from PyQt4.QtCore import pyqtSignal
 
 logfile="lastnotified.log"
 timeformat='%Y%m%d-%H%M%S'
@@ -24,20 +12,20 @@ class lastNotifiedLogger:
         self.tl=None
         self.dt=None
 
-        if os.path.exists(file):
+        if os.path.exists(self.file):
             print('read lnl log')
             self.f=open(file,'r')
             self.line=self.f.readline()
             self.f.close()
             self.dt=datetime.datetime.strptime(self.line.strip(),timeformat)
-            print(self.dt)
+            #print(self.dt)
         else:
             # no file so we should ask anyway
             self.dt=datetime.datetime(1970,01,01,0,0,0)
 
     def timeElapsed(self):
         self.tl= datetime.datetime.now()-self.dt
-        print(self.tl)
+        #print(self.tl)
 
     def excedTimeElpased(self,hours):
         self.timeElapsed()
@@ -48,7 +36,8 @@ class lastNotifiedLogger:
 
     def clear(self):
         print('clear lnl log')
-        os.remove(self.file)
+        if os.path.exists(self.file):
+            os.remove(self.file)
         self.dt=None
         self.tl=None
 
@@ -58,7 +47,7 @@ class lastNotifiedLogger:
         self.dt=datetime.datetime.now()
         self.f.write(datetime.datetime.now().strftime(timeformat))
         self.f.close()
-        print(self.dt)
+        #print(self.dt)
 
 
 if __name__ == "__main__":
